Converted metadata storage to use weak/soft references.
The metadata values are now Objects (instead of primitive types stored in Strings), but the references to those values are stored in SoftReference objects to prevent the metadata store from being the only thing keeping the value object in memory. As a soft reference, the GC can clean up the value in times of memory pressure. The LazyMetadataValue class is smart enough to detect that its internal value has been prematurely garbage collected and will regenerate it on demand.